pyproject.toml
[project]
name = "opensable"
version = "1.1.0"
description = "Your personal AI that actually does things – autonomous, local, and yours forever"
authors = [{name = "IdeoaLabs", email = "contact@ideolabs.ai"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "agent", "agi", "autonomous", "llm", "telegram", "chatbot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "ollama>=0.1.6,<1.0",
    "chromadb>=0.4.22,<0.6",
    "tiktoken>=0.5.2,<1.0",
    "python-dotenv>=1.0.0,<2.0",
    "pydantic>=2.5.0,<3.0",
    "httpx>=0.26.0,<1.0",
    "aiohttp>=3.9.0,<4.0",
    "pyyaml>=6.0.1,<7.0",
    "rich>=13.7.0,<15.0",
    "click>=8.1.7,<9.0",
    "psutil>=5.9.0,<7.0",
    "RestrictedPython>=6.2,<8.0",
    "aiofiles>=23.2.1,<25.0",
    "python-dateutil>=2.8.0,<3.0",
    "PyJWT>=2.8.0,<3.0",
    "py-cpuinfo>=9.0.0,<10.0",
    "cryptography>=41.0,<44.0",
]

[tool.setuptools]
packages = [
    "opensable",
    "opensable.core",
    "opensable.core.tools",
    "opensable.interfaces",
    "opensable.skills",
    "opensable.skills.social",
    "opensable.skills.media",
    "opensable.skills.data",
    "opensable.skills.automation",
    "opensable.skills.trading",
]

[project.optional-dependencies]
core = [
    "aiogram>=3.3.0,<4.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "fake-useragent>=1.4.0",
]

telegram = [
    "aiogram>=3.3.0,<4.0",
    "python-telegram-bot>=20.7,<22.0",
    "telethon>=1.34.0,<2.0",
]

discord = [
    "discord.py>=2.3.2,<3.0",
]

slack = [
    "slack-bolt>=1.18.0,<2.0",
]

web = [
    "fastapi>=0.104.0,<1.0",
    "uvicorn[standard]>=0.24.0,<1.0",
    "websockets>=12.0,<14.0",
]

voice = [
    "faster-whisper>=0.10.0",
    "piper-tts>=1.2.0",
    "pyttsx3>=2.90",
    "elevenlabs>=1.0.0",  # optional cloud TTS (default is local)
]

vision = [
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pillow>=10.1.0",
    "ultralytics>=8.0.200",
    "easyocr>=1.7.0",
]

sync = [
    "websockets>=12.0",
    "cryptography>=41.0.7",
]

automation = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.16.0",
    "fake-useragent>=1.4.0",
]

database = [
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "motor>=3.3.0",
    "redis[asyncio]>=5.0.0",
]

integrations = [
    "discord.py>=2.3.2",
    "slack-bolt>=1.18.0",
    "google-api-python-client>=2.110.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "twilio>=8.10.0",
]

social = [
    "instagrapi>=2.0.0",
    "facebook-sdk>=3.1.0",
    "linkedin-api>=2.2.0",
    "TikTokApi>=6.3.0",
    "python-youtube>=0.9.0",
]

monitoring = [
    "prometheus-client>=0.19.0",
]

cloud = [
    "openai>=1.12.0,<2.0",
    "anthropic>=0.18.0,<1.0",
    "google-genai>=1.0.0",
    "mistralai>=1.0.0",
    "cohere>=5.0.0",
    # deepseek, groq, together, xai use the openai SDK with a custom base_url
]

all = [
    "opensable[core,telegram,discord,slack,web,voice,automation,database,integrations,social,monitoring,cloud]",
]

dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/IdeoaLabs/Open-Sable"
Documentation = "https://github.com/IdeoaLabs/Open-Sable/tree/master/docs"
Repository = "https://github.com/IdeoaLabs/Open-Sable"
Issues = "https://github.com/IdeoaLabs/Open-Sable/issues"

[project.scripts]
opensable = "opensable.__main__:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    "check_system.py",
    "__pycache__",
    "opensable.egg-info",
    "venv",
    "data",
    "logs",
    "whatsapp-bridge",
    "venom-bot",
]

[tool.ruff.lint]
select = ["F"]
ignore = [
    "F811",   # redefinition of unused name (overloads / re-exports)
    "F841",   # local variable assigned but never used (common in try/except patterns)
    "F821",   # undefined name (forward refs / conditional imports)
]

[tool.ruff.lint.per-file-ignores]
# F401 in these files: imports inside try/except for availability checks
"cli.py" = ["F401"]
"demo.py" = ["F401"]
"opensable/core/autonomous_mode.py" = ["F401"]
"opensable/core/browser/browser_engine.py" = ["F401"]
"opensable/core/computer_tools.py" = ["F401"]
"opensable/core/pdf_parser.py" = ["F401"]
"opensable/core/rag.py" = ["F401"]
"opensable/core/voice.py" = ["F401"]
"opensable/interfaces/discord_bot.py" = ["F401"]
"opensable/interfaces/matrix_bot.py" = ["F401"]
"opensable/interfaces/mobile_api.py" = ["F401"]
"opensable/interfaces/whatsapp_bot.py" = ["F401"]
"opensable/skills/advanced_scraper.py" = ["F401"]
"opensable/skills/image_skill.py" = ["F401"]
"opensable/skills/rag_skill.py" = ["F401"]
"examples/*.py" = ["F401", "F841"]
"tests/*.py" = ["F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
